feat(as-4049): rename appeals test data in the business rules
